requirements.txt
fastmcp==0.5.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
authlib==1.2.1
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1
